timeout process

+++ src/main/java/com/example/demo/model/ErrorHandler.java
@@ -4,9 +4,9 @@ public record ErrorHandler(
         String exceptionHandler,
         Integer maximumRedeliveries,
         Double exceptionBackOffMultiplier,
-        Integer redeliveryDelay) {
+        Long redeliveryDelay) {
 
-    public ErrorHandler(String exceptionHandler, Integer maximumRedeliveries, Double exceptionBackOffMultiplier, Integer redeliveryDelay) {
+    public ErrorHandler(String exceptionHandler, Integer maximumRedeliveries, Double exceptionBackOffMultiplier, Long redeliveryDelay) {
         this.exceptionHandler = exceptionHandler != null ? exceptionHandler : "direct:defErrorHandler";
         this.maximumRedeliveries = maximumRedeliveries != null ? maximumRedeliveries : 5;
         this.exceptionBackOffMultiplier = exceptionBackOffMultiplier != null ? exceptionBackOffMultiplier : 2.0;
